goods_srv  商品分类

+++ goods_srv/utils/public.go
@@ -27,7 +27,7 @@ func GetMenu(parentId uint32) []*proto.CategoryInfoResponse {
 				Id:       item.ID,
 				ParentCategory: item.ParentCategoryId,
 				Name:     item.Name,
-				Level:    int32(item.Level),
+				Level:    item.Level,
 				IsTab:    item.IsTab,
 			}
 			node.SubCategorys = child
